7196 Mouse pointer no longer visible in HMD mode
The issue was caused by creation of additional instances of 'Cursor::Manager' in plugins, which happened because
Cursor::Manager's code is inside static library so every module was getting its own copy of it. As the result cursor-related settings from Application never reached plugins.

The fix is based on converting 'Cursor::Manager' into 'Dependency' to avoid additional instances creation in plugins