Fix .fst texdir not being followed
Although the texdir was being acknowledged and used as the
_textureBaseURL inside of the Geometry* classes, it was being
overwritten in code meant to handle redirects. Basically, when a
geometry resource request is redirected (via ATP, HTTP, etc.), we needed
to update the _textureBaseURL to take the new location into account.
Previously we were overwriting the _textureBaseURL all the time, even
when not being redirected, but this updates it to only be overwritten
when the request is redirected.

There is at least 1 known case that this does not handle: a .fst with
its `texdir` set, that points at an fbx that gets redirected.