Fix KTX requests thrashing if a mip level can't be assigned
When downloading a mip level and assigning it to the underlying
resource, we weren't checking to see if the assignment was successful.
If it was unsuccessful, we would queue up a request for the next mip
which would end up being a request for the same mip that was just
attempted. This change fixes this problem by checking to see if the
texture's available mip level is at least <= the assigned mip level.