Crash/Deadlock fix: Web3DOverlay could be destroyed on wrong thread.
When script calls Entities.getChildrenIDs*() it is possible to deadlock the main thread, and also invoke
the destructor of Entities and Overlays, which is very NOT thread safe.

The fix is to use a pattern already in use in several places in our codebase.
Use the custom deleter parameter of std::shared_ptr to call deleteLater() instead of destroying the object in place.
This allows any thread to use shared_ptrs of SpatiallyNestables without fear.