Fix baking bug with single-part FBX + multiple materials
In FBXBaker we were not packing in material IDs if there was only 1 part
in the mesh. This was unintentionally making the assumption that there
was only 1 material at index 0 - if there are more materials, the material index
that the part references may not be 0. The change ensures that we pack
material IDs if there are multiple parts or if the sole part uses a
material ID other than 0 (0 is the default).