[Case 7049] Fixes superfluous fields issue with Particle Effects (details below).
When creating or selecting an object entityProperties.html's properties-list is
updated to have and entity type specific class.  This class is used within
edit-style.css to specify rules for the Properties Tab when a certain type of
entity is currently selected.

This resolves the issue with duplicate/misleading color fields for Particle
Effects by adding .ParticleEffectMenu rules to edit-style.css to control the
visible sections.

When the Properties Tab is in the ParticleEffectMenu mode only the following
sections are shown (listed as ordered):
    General Section
    Collision Section
    Physical Section
    Spatial Section
    Behavior Section

Changes Committed:
    modified:   scripts/system/html/css/edit-style.css
    modified:   scripts/system/html/js/entityProperties.js