Improve stability for HMD third person avatars
Previously there was a hack where the sensor to world matrix was NOT updated in HMD mode if the camera was third person.
The reason the hack existed was to make third person camera less nauseating for users.
However this had a major drawbacks.
It led to avatar hip spasms when the user turned around in the tracked volume, while in third person camera.

To fix this issue properly, the hack which prevented updating the sensor to world matrix was removed, and now the third person camera is now computed in sensor space relative to the hmd.