Much more stable hips IK target.
Target is now estimated in sensor space from the head, which is very deterministic and not prone to feedback from the IK system.
Previous the hip was estimated from accumulated IK error deltas, which was not stable, as the error would tend to accumulate if the IK targets could not be reached.