[Case 4315] Fixes color picker issue in HMD mode (details below).
Previously when in HMD mode, using the color picker to select a color via the
gradient or hue areas would result in the page scrolling making it difficult to
select the color or hue desired.

This resolves the issue by turning off touch actions for the elements
of the color picker that should have it when using the color picker.

Tested in HMD & Desktop mode.

Changes Committed:
    modified:   scripts/system/html/css/colpick.css