Bug fix for avatar neck cauterization
The basic dual-quaternion skinning algorithm does not handle non-rigid transformations like scale well.
Because we only use scaling for head cauterization, we special case this by passing in a cauterization factor,
as well as a cauterization position to the vertex shader.  If a vertex is flagged as cauterized, we slam it to equal the cauterization position.
Although, not as smooth as the previous method, it seems to work well enough on the avatar's I've tested.