Revert "FBX node IDs aren't alphanumerically ordered per logical structure"
This reverts commit a7ec4501e65396d4c7bf2316dd73188cea7a3c7d.

Because remainingModels is a QSet, the order is not guaranteed.
Therefore the same code iterating over the same items will sometimes have a different ordering.
See docs for QSet, http://doc.qt.io/qt-5/qset.html
This was bug was causing scrambled avatars, because both the transmitter and receiver of the AvatarData packets
make the strong assumption that the joint orders are same.  When they are not the avatar's appear scrambled.