Remove unused entity server backups
The domain server now handles entity backups so this code is no longer
used.

This also fixes a bug where, if the entity server went down while
writing the persist file, it would leave a lock file behind. When
starting up the next time it would try to recover from the last backup.
Because the entity server no longer gets any backup rules, and the way
it searches for backups, it would end up finding the lock file and
treating it as the most recent backup. This meant that the
models.json.gz file was replaced with an empty file.