Remove logging messages for behaviors that are expected.
Attached entities with scripts are now being sold on the marketplace
(rWatch) and are now being worn by various users.  Because of this, it's
now common for the 'removing entity scripts' case to be hit in code for
every update, which causes debug log messages every 60 seconds for each
entity entity with a script.

Repro: Have an avatar wear rWatch.  With a second instance of interface,
turn on verbose logging while the first avatar is local.  See debug logging
60 times a second.

Fix: Given attached entities with scripts are happening, it's not an
exceptional case and should probably not be logged.