Remove UndoStack API
edit.js shares a global undo stack through the UndoStack API. Because
the stack is globally shared, disabling the feature while create is
disabled doesn't make a lot of sense because it might disabling the
ability to undo/redo with another feature. Because UndoStack is only
used by edit.js, UndoStack is being removed in favor of a local undo
stack implemented within edit.js.