In CCD IK chains, remove the LeftHand and RightHand constraint completely.
When used in combination with the elbow pole vector, this was causing the hand to never properly reach its rotation target.
It was also having side effects causing the elbow to never straighten properly to reach its position target.

There is possibly a bug in how this tip constraint is enforced and I've verified that there are no ill effects in removing it.