Fix for bad static_pointer_cast on exit.
Identified by Address Sanatizer on Linux.

On exit ~Applicaiton calls AvatarManager::deleteAllAvatars.  This iterates over all avatars remaining in the _avatarHash.
Typically this includes the MyAvatar instance.  In this case the previous code would attempt to static cast an AvatarData pointer to an instance
of OtherAvatar.  This might be harmless, but could also lead hard to diagnose issues such as memory corruption.