Fixes for OSX
App Translocation makes it nearly impossible to find the interface
executable from the running server-console, specifically in downloaded
builds launched from the Finder.  To enable OSX detection and and
launch of the interface, the code has been changed to:
* Assume interface is installed on OSX as we don't produce a server-only
build.
* Launch the interface by using 'open' with the appid.
NOTE: This may launch the wrong version of the app if
multiple instances of the app are installed, but this
is the best we can do for now.  For most users, the
behavior should be as expected.

Second, polling was happening even if interface couldn't be
detected if it was previously enabled.