Bug fix for rotation culling in AvatarData::toByteArray()
The commit 5a0de0f103c inadvertently introduced a regression in how the AvatarMixer and clients encode AvatarData packets.
When encoding packets that with the cullSmallChanges flag set to true, we only include rotations that have changed significantly from the previous packet.
A bug was introduced that incorrectly preformed the comparison to detect changed rotations, this needs to handle the case when
two quaternions are almost equal but have negative signs.  Previously, this was preformed with an absolute value after the 4-component dot product.
In commit 5a0de0f103c, this absolute value was removed, causing the culling to happen much more frequently.

This PR re-introduces that absolute value.