AnimDebugDraw: Prevent multi-threaded access to shared map collection.
It was possible for the render thread to iterate over a map while the main thread mutates the same map.
This could have led to memory corruption and crashes.

To fix this we make a copy of the collection and pass the copy to the lambda that executes on the render thread.