Head and Hand offsets for Vive Trackers now function as expected
In the Settings > Controller... > Calibration menu, when using head or hand trackers
The provided Y and Z offsets should function correctly now.  This gives you the ability to adjust
the offsets of the pucks to the hands/head.  For example: this allows you to wear the hand pucks on your forearms
rather then the backs of your palms.