Use QSaveFile when persisting domain settings to disk
Prior to this change the `DomainServerSettingsManager::persistToFile()`
method wrote directly to the settings file. It also did limited error
checking. These limitations could lead to a situation where a crashed
domain-server process, a file system backup, a settings file copy, or a
hardware error could corrupt the setting file.

This change swaps the `QFile` class for `QSaveFile` and uses its atomic
saving features, which writes the changes to the settings file to a file
in the same directory as the settings file and then does a rename to
replace the original. If the rename, or any of the file operations, fail
the original settings file remains in place.