Bug fix for avatars that have mesh/joint name duplicates
This change has two parts.  First we do a better job of identifying joints vs meshes in the FBXSerializer.
Second we use this information to better handle duplicate names in AnimSkeleton.
When a duplicate is detected, allow joints to override meshes.
This will ensure that when joints are looked up by name we will not return a mesh object with the same name by mistake.