Add prebuild metrics logging for CI system
Metric about the times various parts of the build take are great to
have. This change takes us a step in that direction by adding metrics to
the prebuild step of the build process. These metrics are off by
default; use the `--ci-build` option of `prebuild.py` to enable them.