Fixed a race condition that sometimes caused the main menus not to appear.
When Interface starts, it first calls pauseUntilLoginDetermined(), and later resumeAfterLoginDialogActionTaken().
But on rare occasions these functions are called in the reverse order, and this caused Interface to remain
in the "paused" state. Now we check for this case, and abort pauseUntilLoginDetermined() if it happens.

This has only happened to me when running Interface immediately after rebuilding it (in Release mode).