Fixed a race condition related to the keyboard when starting Interface.
These two things need to happen in the following order:
1. Initialize the input plugins (including the keyboard)
2. Start the scripts

That's because the scripts try to use the keyboard (e.g., in edit.js), and if it's not ready yet
then they fail, and then Interface doesn't work right (e.g., the Create button doesn't work).

In order to ensure that these things happen in the correct order, we must make sure that
resumeAfterLoginDialogActionTaken() (which starts the scripts) is called only after everything
else in the Interface constructor has finished. Usually this happens correctly, but occasionally
resumeAfterLoginDialogActionTaken() is called too soon. This commit makes sure that even in that
case, we'll postpone calling resumeAfterLoginDialogActionTaken() until the Interface constructor
has finished.