Fixed Safe Landing interaction with Local entities.
Local entities are collisionless, so they shouldn't affect Safe Landing since it specifies that it only wants
to find COLLIDABLE entities. However, due to the requirement to support legacy behavior, picks for COLLIDABLE
entities *do* intersect Local entities. In order to prevent this, we have to explicitly request only intersections
with Domain or Avatar entities.

For more information about this, see this Pull Request:
https://github.com/highfidelity/hifi/pull/15282