Case 21726 - Domain lost recent entity edits upon restart
When an entity server starts up, grabs the version of the models.json
file locally, and then queries the domain server for its copy of the
models.json file...iff the domain server version is newer.

There was a bug in this process in that the comparison was made between
the wrong version, specifically the 'file format version' which doesn't
change unless there was a protocol change...and not the data version,
which increments every time a change is made to a domain.

Therefore, the version of the models.json on the domain server was never
downloaded to the entity server, even when it was newer.

It would be newer if the entity server assignment was moved to a machine
with an old version of the models.json file, which was in fact the case
on distributed3 during this period of time.