Fix domain baker not finishing (threading, material key)
This addresses 2 cases where the domain baker would not finish

1. For material baking, we were storing a value in
_entitiesNeedingRewrite with a different key than what
we use when we later remove it from _entitiesNeedingRewrite

2. The domain baker runs on a worker thread. When starting a
baker for an item in the domain, we do an invokeMethod,
which will happen synchronously if the baker ends up on the
same thread as the thread that the domain baker is on.
This can cause issues if the baker finishes immediately.
The case I saw was a local model url that immediately failed,
and finished before return from the invokeMethod in the domain baker .