BUGZ-87 - loading/physics stuck when entering domain
A race condition was occurring when multiple entities using
the same models were handed over to physics around the same time.
The first model is passed to a worker thread to turn it into a shape
and subsequent models note they are the same and don't create a shape
or worker thread.

The physics system was ignoring these, never marking them as ready
for physics.