Fix blender::run crash
Pass a shared_ptr of the HFMModel to the Blender.  This will prevent the HFMModel
from being destroyed on the main thread if an Avatar changes their skeletonModelURL.

Also the _blendshapeOffset hash in Model has been eliminated, it was not necessary and
was also a source of data races.

The body of Blender::run has been updated slightly to reduce the number of allocation necessary for temporary QVectors.