BUGZ-223 - Physics was not starting in TheSpot-dev This was caused by two issues involving the SafeLanding system: * Race condition on domain load The domain is hard-reset twice during typical start, first after login completes, and second, when ice/domain renegotiation occurs Entities added during the first reset caused an add to the safe landing system which, because those adds were queued, possibly happened after the second hard reset.
Second, Safe landing checks whether physics is ready on entities before
letting physics start.  Physics is ready when the entities have a
physics shape associated with them if they're static meshes.  The
physics shape is only generated if the entity is in Region 1 or 2.
Safe landing, however, was not checking the region of the entity
and was therefore waiting on entities that were not in region 1 or 2