Use version from Interface in macOS HQ Launcher
Prior to this change the macOS HQ Launcher used the version from the
settings files. Pull this information in the settings file could
result in a situation where the setting file was updated, but interface
wasn't. When the launcher ran a second time it would check the settings
file and think Interface was updated even though it wasn't. This change
resolves the issue by asking interface for its version directly.

In the event that Interface doesn't response with a version number HQ
Launcher knows how to interoperate HQ Launcher will fall back to using the
settings file.

Known Issues
---

1) Interface reports different version number formats depending
    on how its built. A stable interface for getting the build version
    from interface should be added.