Never cache the tags API in macOS's HQ Launcher
Prior to this change we used the default session while downloading the
latest tags API response from Thunder. This left open the possibility that
NSURLSession could cache the response, even though the server asked it
not to; NSURLSession can be pretty aggressive about its caching.

This change ensure no caching of the API is occurring my using an
ephemeral session when calling the tags API.