Enable QtWebEngine to store its cache on disk
In the past, QtWebEngine stored the cache on disk automatically. But apparently due to a recent
change to Qt, this has changed: a Web Engine Profile that is constructed without a Storage Name is placed
in "Off the Record" mode, which means that the cache is stored only in memory. Therefore, we must assign
a Storage Name and disable "Off the Record" mode.