Fix cmd-leftclick not working correctly on Mac OS
On Mac OS, Cmd+LeftClick is treated as a RightClick (more specifically, it seems to
be Cmd+RightClick without the modifier being dropped). Starting in Qt 5.12, only
on Mac, the MouseButtonRelease event for these mouse presses are sent to the top
level QWidgetWindow, but are not propagated further. The change here gets around
this problem by capturing these pseudo-RightClicks, and re-emitting them as
"pure" RightClicks.