Check macOS Launcher's zip size before unzipping
The Launcher can download an error page rather than the full Launcher.
When this happens, the error page file is passed directly to unzip,
which promptly fails with the following message:

  End-of-central-directory signature not found.  Either this file is not
  a zipfile, or it constitutes one disk of a multi-part archive.  In the
  latter case the central directory and zipfile comment will be found on
  the last disk(s) of this archive.
  unzip:  cannot find zipfile directory in one of NewHQLauncher.zip or
          NewHQLauncher.zip.zip, and cannot find NewHQLauncher.zip.ZIP, period.

This change improves the error message by checking the size of the
Launcher download and failing if the Launcher is smaller than the
smallest Launcher we've released. The smallest Launcher we've had is
308KB, so I picked 250KB as the limit because it's a nice rounder number
below that.