Fix for avatars that are stuck in idle pose for observers
The problem can occur because the default constructor for the NetworkAnimState does not initialize the blendTime member.
This can cause the avatar to remain stuck in an idle pose for observers when blendTime happens to be randomly initialzied to a large negative floating point value.
To fix this we explicitly initialize it to FLT_MAX.

Additionally, a debug print has been added to help diagnose this issue if it occurs again.