Remove `this` and connection from lambda expression
Clang in Xcode 11 is more aggressive about its warnings. This resolves
one of the issue Clang noticed in order to support building w/Xcode 11.

I first attempted to remove `this` completes, but VS complained that
`QObject::disconnect` was being called without `this`. This change
satisfies both Clang and VS.

A second attempt added this back and then called `disconnect` directly.
This also generated a Clang warning.

This version resolves the issue by removing the error handler. This also
resolves a memory leak caused by the error handler.