Speculative fix for infinite loop in SafeLanding
This PR containts two fixes.

1) Tightening up the locks in the SafeLanding class to prevent race conditions.
By inspection there are cases where variables are read and modified outside of locks.

2) A check to help prevent an infinite loop in the _sequenceNumber std::distance calculation.
This is the main issue, in some cases backtrace is reporting the main thread as deadlocked.
The stacktrace points to an issue with the std::distance() calculation.
If the EntityQueryInitialResultsComplete is delayed signficantly, there could be a rare
case where the _sequenceNumber map grows large enough for the wraparound less then operator
will no longer function correctly. This will cause the std::distance calculation to never complete.
I've added a guard to prevent this from happening and some logs to help diagnose this issue in the future.