Change asset URLs to Athena S3 bucket.
This uses a variable that can be overriden with an environment variable.
This makes it easy for the community to create alternate asset hosts.

Due to vcpkg using a clean environment, variables can't be easily passed
to cmake scripts run by vcpkg. This was worked around by writing the data
that needs to be passed into temporary files and reading it from the
vcpkg ports.