Redo graphics presets
* Add "battery saver" mode, which is the only one that has "economical"
  rendering. This mode will never be selected automatically and must be
  chosen by the user.
* Change all other modes to realtime.

Reasoning:

Whenever interface crashes, it sets the graphics settings to low. This not
only has low detail, but also limited framerate. This may create the impression
that there's something seriously wrong with the rendering engine.

Now the interface will always have the best framerate, except when the user
wishes otherwise.