Zone selector sorting vs uppercases
This makes the sorting of the Zone in the Zone selector
with the uppercase grouped with the lowercase for a same letter
but uppercase always first.
(Note: Accented char are considered as a different letter. We would need to support exception cases for all the existing charsets. These are only entities descriptors, not really for literature.)