Selection Tool "unlit"
This commit includes 2 parts:
1- Re-adjustment in the code to stay aligned with axis/colors/Pitch-Yaw-Roll 
It seems that the Overlay 3dCircle was in fast patch to work while the Gizmo Ring is aligned for Axix versus Pitch-Yaw-Roll.
The modification has been done differently and is in fact a simpler solution.

2- Add a Material local entities parented to each tool entity to make the Selection tool "unlit"
This addresses Issue #709 Enhancement: Make create translate "widget" unlit
BUT there is a known issue: The Gizmo Ring are not supporting the Material currently. (an issue has been open for that see #744)
The Rotation rings are then not "unlit", but everything in the code is ready to support it when it will be done.
Gizmo Ring having practically designed for that specific usage, I think we should keep using them.
Other alternatives like using an image or a model would imply to considerable modifications. 
We would lose the angular representation (that can't be done with image or model). 
The remaining "unlit" issue is very low severity, the widget is still fully functional. So I think it can wait for the support of Material.