Fix a bug about tablet stuck in Landscape
Fix a bug about tablet stuck in Landscape
when the create app was open in desktop and you put your VR Headset on, 
then next time you open the tablet it was in landscape mode. (because a refresh was happening after the landscape set to false at the closure)
Now the condition is geared to deal with the situation.