HMD: Make the 2 buttons visible without scrolling
HMD: Make the 2 buttons visible without scrolling

In HMD, the 2 buttons:
- Open This Domain's Asset Server
- Import Entities (.json)
were not visible without scrolling. people would not even notice their presence
This fix adjust the spacing and the margin to have the entities button on 2 rows when the UI is in landscape mode
this way the button are visible.