Set OpenGL_GL_PREFERENCE to LEGACY
GLVND appears to break the entire UI on ATI video cards.

It was set to GLVND due to an understanding that this was a no-op setting, but it turns
out the documentation in OpenGL_GL_PREFERENCE is confusing, and it looks like LEGACY
is the setting that was being used after all.

An additional point of interest is that debian/rules contained OpenGL_GL_PREFERENCE=GLVND,
and this was done back in the High Fidelity days. It's uncertain whether this was correct
or not, though.