"Teleport To Selected Entities" back to Selection
This returns "Teleport To Selected Entities" back to the "Selection..." menu.
After having tried a certain time what was suggested (having it under "Actions..." menu)
This is clearly easy to confuse it with "Move Selected Entities to Avatar"
"Teleport To Selected Entities" is about position your-self to see what you have "Selected" (which make this a Selection Tools or an assistant to it)
while "Move Selected Entities to Avatar" is clearly an action. as it moves the entities to a new position.